# Add flatten_variable